refactor(osn/api): split into pure createApp factory + Bun dev entry
Phase 1 of the Cloudflare Workers migration — pure refactor, zero
behaviour change; everything still runs on Bun and all 571 tests stay
green.

- src/app.ts: `createApp(deps)` holds the Elysia route composition
  verbatim, taking an explicit AppDeps struct and never reading
  process.env.
- src/local.ts: `buildAppDeps()` does all env-driven wiring (JWT key
  pair, session-IP pepper, Redis stores + rate limiters, email transport,
  the shared ManagedRuntime built ONCE) and `startBunServer()` keeps the
  listen + ephemeral-key warning + outbound ARC rotation + erasure sweeper.
- src/index.ts: thin Bun entry tests import — builds deps, calls
  createApp, still exports `app`, still listens conditionally on NODE_ENV.

Redis/ioredis, observability, and the Workers fetch entry are left for
later phases.
